[finishes #161742752] Create Endpoint that makes parcel delivery orders
In the test_user.py file create mock data for test use, and tests for the
parcel endpoint that test, whether the user has submitted an empty post,
or a submitted post is missing values, or whether the user is missing
some fields and a final test to create a delivery order.

In the modals.py file create a function that will create the parcel
delivery order and save it temporarily in a list.

In the user.py file, create a route of POST method with a function
that is able to check whether an empty delivery order has been made,
check whether the delivery order contains empty values and can point
out which field is missing a value, also check whether some fields are
missing and finally make a delivery order.

In the errors.py file create a helper function which is passed a numerical
status code and an error messages as arguments and is able to change the
status code to text and return both the text version of the status code and
the message in user friendly jsonified format.