Change to using pytest from unittest
Create a config file that will contain all fixtures. In here,
write a fixture that will create a flask app, set up a testing
client and run it to test endpoints and finally close it.

Change the tests in the test_user file to work with pytest fixture.
move the mock data to the config file