[finishes #161742822] Create endpoint to fetch a delivery order by its id
Add an endpoint of method get that takes a parameter of type integer
in the url and returns a message if no delivery orders exist, or returns
a message containing the id and the details of the order id, and if the
id doesnt exist in the records it returns a message.

Write tests for this endpoint for the conditions when an id exists and
when an id doesnt exist.