Add the user id field
In the modals file add the field of user id.

Add the user_id field to the test mock data in the conftest file.

Add a test to the test_user file that checks if the same user_id
already exists.

In the user file on the endpoint that creates a parcel delivery
order, add a condition to check if a user id exists and fails with
a message stating the name of the existing user_id.
if the condition fails, a new parcel delivery order can be created.

The user id should be used to uniquely identify a delivery order,
so that its easy and safer to cancel delivery orders.

[feature #161742752] adds the user_id
[feature #161742636] adds the user_id
[feature #161742822] adds the user_id