[finishes #161743013] Create endpoint to fetch all records by userid
Add an endpoint that takes a parameter userid and returns all delivery
orders that contain the same userid.
If no delivery orders exist return a messsage.
if there are no delivery orders by the userid return a statement.

Add tests that check when a user id exists and when a userid doesn't
exist.