[finishes #161742877 ] Create endpoint to make a parcel delivery order
Create an endpoint of method put that takes a parcelid parameter and
changes the status of the delivery order to canceled. Endpoint should
return a message if no delivery orders exist or if that particular order
doesn't exist.

Write tests for the endpoint to cancel a delivery order.
Change the name of the parcel's id from order_id to parcel_id

[edits #161743013] changes order_id to parcel_id
[edits #161742822] changes order_id to parcel_id
[edits #161742636] changes order_id to parcel_id
[edits #161742752] changes order_id to parcel_id